pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ngsp-clinical"
version = "0.1.0"
description = "Neural-Guided Semantic Proxy for clinical-trial privacy: a research prototype composing Safe Harbor, Gemma-4-routed query synthesis, and calibrated differential privacy in front of the Anthropic API."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "NGSP Clinical Research Prototype" }]
dependencies = [
    "torch>=2.2",
    "transformers>=4.44",
    "accelerate>=0.33",
    "huggingface_hub>=0.24",
    "openai>=1.0",
    "opacus>=1.5",
    "sentence-transformers>=3.0",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "rapidfuzz>=3.9",
    "datasets>=2.20",
    "matplotlib>=3.8",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]
quant = [
    "bitsandbytes>=0.43",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["ngsp*", "attacks*", "data*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "slow: heavy integration tests that load Gemma or hit the Anthropic API (opt in via NGSP_RUN_HEAVY=1)",
]

[tool.ruff]
line-length = 100
target-version = "py310"
